+++ src/utils/settings.py
@@ -134,6 +134,11 @@ def sys_argv_errors():
         err_msg = "Illegal (non-console) comma character ('" + sys.argv[i] + "')."
         print print_critical_msg(err_msg)
         raise SystemExit()
+    # Check for potentially miswritten (illegal '=') short option.
+    if re.search(r"\A-\w=.+", sys.argv[i]):
+        err_msg = "Potentially miswritten (illegal '=') short option detected ('" + sys.argv[i] + "')."
+        print print_critical_msg(err_msg)
+        raise SystemExit()
 
 # argv checks
 sys_argv_checks()
@@ -146,7 +151,7 @@ APPLICATION = "commix"
 DESCRIPTION_FULL = "Automated All-in-One OS Command Injection and Exploitation Tool"
 DESCRIPTION = "The command injection exploiter"
 AUTHOR  = "Anastasios Stasinopoulos"
-VERSION_NUM = "2.7.17"
+VERSION_NUM = "2.7.18"
 STABLE_VERSION = False
 if STABLE_VERSION:
   VERSION = "v" + VERSION_NUM[:3] + "-stable"